HaskellWorks.Data.Aeson

Documentation

newtype JsonEndo a

Constructors

JsonEndo

Fields

unJsonEndo :: [a] -> [a]

newtype WithJsonKeyValues a

For use with language extension DerivingVia. This derivation provides a ToJSON instance that delegates to the ToJsonKeyValues instance.

For example:

newtype MyType = MyType
  { myField :: Text
  } deriving J.ToJSON via WithJsonKeyValues MyType

Constructors

WithJsonKeyValues

Fields

unWithJsonKeyValues :: a

class ToJsonKeyValues a where

Generate key values from a value of a type. This can be used in conjunction with WithJsonKeyValues to define a ToJSON instance without having to implement both toJSON and toEncoding, thereby reducing boilerplate.

For example:

instance ToJsonEncoding MyType where
  toJsonEncoding sv =
    [ "my_field" .!= sv ^. #myField
    ]

Methods

toJsonKeyValues :: (KeyValue kv, Monoid kv) => a -> [kv]

objectWithoutNulls :: [Pair] -> Value

readJson :: Read a => String -> String -> Parser a

objectEndo :: [JsonEndo Pair] -> Value

Same as object except used in combination with .?= and .!= instead of .=.

For example:

toJSON o = objectEndo
  [ "mandatory" .!= o ^. the "mandatory"
  , "optional"  .?= o ^. the "optional"
  ]

(.?=) :: (KeyValue p, ToJSON v, Monoid p) => Key -> Maybe v -> p infixr 7

Render optional fields as missing in JSON output.

(.!=) :: (KeyValue kv, ToJSON v) => Key -> v -> kv infixr 7

Same as .=, but with lower precedence to work well with lens.